feat: Update guacd version default to 1.5.2
Guacamole 1.5.0 introduced the ability to connect to SSH servers using
elliptic-curve keys (in addition to RSA), which is a requirement for
connecting to sshd servers that have phased out RSA key support.

Since then, two interim release have followed to address issues
reported with the Docker images.

Update the default for HASTEXO_GUACD_VERSION from 1.4.0 to 1.5.2,
and also update the README and changelog.

References:
https://guacamole.apache.org/releases/
https://guacamole.apache.org/releases/1.5.0/#release-notes
https://guacamole.apache.org/releases/1.5.2/#release-notes