fix: Fix installing the hastexo XBlock for version >=7.10.1
In the latest changes to the XBlock(v7.10.1) we dropped the "-e ."
in `requirements/base.txt`; we still need to install the code in
the current directory for this plugin so we need to add the "."
in the Dockerfile.

For backward compatibility, install the code still in editable mode
within the Quince release.